Set an expectation about cluster creation time
When creating a cluster using the example command provided, the experience in the CLI does not communicate the expected time needed for cluster creation. By providing this "note", we can set an expectation that this _could_ take up to 30 minutes (this time duration was informed by a conversation with the developers on the team).