Add location of ca cert to documentation
This doc currently leads users to another page which has not any more information about how to locate the ca cert.
If you look at the "advanced troubleshooting with the bosh cli" it starts much the same, with sshing onto the opsman, but it actually
tells you were to locate the certificate. If I followed these directions, I would have to a) follow link, b) follow second link, c) Realize I cannot curl endpoint for certificate until I have a bearer token for opsman. d) have to look up more documentation on how to get a bearer token, and re look up various credentials.

If you are already on the opsman there is no reason for this.