Change StorageClass creation process
When you create a StorageClass on vSphere, there is no need to specify:

- datastore
- diskformat
- fstype

The PKS tile is configured with defaults that you can use for the datastore, and kubectl contains defaults for the [others](https://vmware.github.io/vsphere-storage-for-kubernetes/documentation/storageclass.html). The person that configures the storageclass may not have any idea what a datastore is, or what it _might_ be named. Ditto for diskformat and fstype.

This change removes the over-complicated example.